requirements.txt
python-dotenv==1.0.0pip install -r requirements.txt